feat: created rick and morty select container with props
