+++ src/contents/rick-and-morty-select/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./RickAndMortySelect";
+
+export * from "./RickAndMortySelectProps";
